refactor approaches
