+++ modules/text/module_text_llm/module_text_llm/chain_of_thought_approach/generate_suggestions.py
@@ -1,35 +1,25 @@
-from typing import List, Optional, Sequence
+from typing import List, Optional
 from pydantic import BaseModel, Field
 
 from athena import emit_meta
 from athena.text import Exercise, Submission, Feedback
 from athena.logger import logger
+
+from module_text_llm.chain_of_thought_approach import ChainOfThoughtConfig
+
 from llm_core.utils.llm_utils import (
     get_chat_prompt_with_formatting_instructions, 
     check_prompt_length_and_omit_features_if_necessary, 
     num_tokens_from_prompt,
 )
 from llm_core.utils.predict_and_parse import predict_and_parse
-from module_text_llm.config import BasicApproachConfig
+
 from module_text_llm.helpers.utils import add_sentence_numbers, get_index_range_from_line_range, format_grading_instructions
-
-class FeedbackModel(BaseModel):
-    title: str = Field(description="Very short title, i.e. feedback category or similar", example="Logic Error")
-    description: str = Field(description="Feedback description")
-    line_start: Optional[int] = Field(description="Referenced line number start, or empty if unreferenced")
-    line_end: Optional[int] = Field(description="Referenced line number end, or empty if unreferenced")
-    credits: float = Field(0.0, description="Number of points received/deducted")
-    grading_instruction_id: Optional[int] = Field(
-        description="ID of the grading instruction that was used to generate this feedback, or empty if no grading instruction was used"
-    )
+from module_text_llm.chain_of_thought_approach.prompt_thinking import InitialAssessmentModel
+from module_text_llm.chain_of_thought_approach.prompt_generate_feedback import AssessmentModel
 
 
-class AssessmentModel(BaseModel):
-    """Collection of feedbacks making up an assessment"""
-    
-    feedbacks: List[FeedbackModel] = Field(description="Assessment feedbacks")
-
-async def generate_suggestions(exercise: Exercise, submission: Submission, config: BasicApproachConfig, debug: bool) -> List[Feedback]:
+async def generate_suggestions(exercise: Exercise, submission: Submission, config: ChainOfThoughtConfig, debug: bool) -> List[Feedback]:
     model = config.model.get_model()  # type: ignore[attr-defined]
 
     prompt_input = {
@@ -43,10 +33,12 @@ async def generate_suggestions(exercise: Exercise, submission: Submission, confi
 
     chat_prompt = get_chat_prompt_with_formatting_instructions(
         model=model, 
-        system_message=config.generate_suggestions_prompt.system_message, 
-        human_message=config.generate_suggestions_prompt.human_message, 
-        pydantic_object=AssessmentModel
+        system_message=config.thikning_prompt.system_message, 
+        human_message=config.thikning_prompt.human_message, 
+        pydantic_object=InitialAssessmentModel
     )
+    
+
 
     # Check if the prompt is too long and omit features if necessary (in order of importance)
     omittable_features = ["example_solution", "problem_statement", "grading_instructions"]
@@ -66,22 +58,48 @@ async def generate_suggestions(exercise: Exercise, submission: Submission, confi
             emit_meta("error", f"Input too long {num_tokens_from_prompt(chat_prompt, prompt_input)} > {config.max_input_tokens}")
         return []
 
-    result = await predict_and_parse(
+    initial_result = await predict_and_parse(
         model=model, 
         chat_prompt=chat_prompt, 
         prompt_input=prompt_input, 
-        pydantic_object=AssessmentModel,
+        pydantic_object=InitialAssessmentModel,
         tags=[
             f"exercise-{exercise.id}",
             f"submission-{submission.id}",
-        ]
+        ],
+        use_function_calling=True
     )
 
+    second_prompt_input = {
+        "answer" : initial_result.dict(),
+        "submission": add_sentence_numbers(submission.text)
+
+    }
+    
+    second_chat_prompt = get_chat_prompt_with_formatting_instructions(     
+        model=model, 
+        system_message=config.generate_suggestions_prompt.second_system_message, 
+        human_message=config.generate_suggestions_prompt.answer_message, 
+        pydantic_object=AssessmentModel)
+    
+    result = await predict_and_parse(
+    model=model, 
+    chat_prompt=second_chat_prompt, 
+    prompt_input=second_prompt_input, 
+    pydantic_object=AssessmentModel,
+    tags=[
+        f"exercise-{exercise.id}",
+        f"submission-{submission.id}",
+    ],
+        use_function_calling=True
+    )
+        
     if debug:
         emit_meta("generate_suggestions", {
             "prompt": chat_prompt.format(**prompt_input),
             "result": result.dict() if result is not None else None
         })
+
 
     if result is None:
         return []
