+++ modules/text/module_text_llm/module_text_llm/chain_of_thought_approach/prompt_thinking.py
@@ -1,3 +1,6 @@
+from pydantic import BaseModel, Field
+from typing import Optional, List
+
 system_message = """
 You are a grading assistant at a prestrigious university tasked with grading student submissions for text exercises.
 You goal is to be as helpful as possible to the student while providing constructive feedback without revealing the solution.
@@ -27,3 +30,30 @@ Student\'s submission to grade (with sentence numbers <number>: <sentence>):
 {submission}
 \"\"\"\
 """
+
+# Input Prompt
+class ThinkingPrompt(BaseModel):
+    """\
+Features available: **{problem_statement}**, **{example_solution}**, **{grading_instructions}**, **{max_points}**, **{bonus_points}**, **{submission}**
+
+_Note: **{problem_statement}**, **{example_solution}**, or **{grading_instructions}** might be omitted if the input is too long._\
+"""
+    system_message: str = Field(default=system_message,
+                                description="Message for priming AI behavior and instructing it what to do.")
+    human_message: str = Field(default=human_message,
+                               description="Message from a human. The input on which the AI is supposed to act.")
+  
+# Output Object
+class InitialAssessment(BaseModel):
+    title: str = Field(description="Very short title, i.e. feedback category or similar", example="Logic Error")
+    description: str = Field(description="Feedback description")
+    line_start: Optional[int] = Field(description="Referenced line number start, or empty if unreferenced")
+    line_end: Optional[int] = Field(description="Referenced line number end, or empty if unreferenced")
+    credits: float = Field(0.0, description="Number of points received/deducted")
+    reasoning: str = Field(description="Reasoning why the feedback was given")
+    impprovment_suggestion: str = Field(description="Suggestion for improvement for the student")
+
+class InitialAssessmentModel(BaseModel):
+    """Collection of feedbacks making up an assessment"""
+    
+    feedbacks: List[InitialAssessment] = Field(description="Assessment feedbacks")
